Fixes tests

+++ test/tests/Sensor/SensorAccelerometerTest.cpp
@@ -30,34 +30,34 @@ TEST_CASE( "SensorAccelerometer::begin -- pitch", "[SensorAccelerometer][Sensor]
         mock.setReturnValues(Vec3 { 0, -GRAVITY, 0 }, Vec3 { 0, 0, 0}, true);
         accel.begin();
         
-        REQUIRE( accel.getIncline() == 0 );
+        REQUIRE( accel.getIncline().toFloat() == 0 );
     }
 
     SECTION("Accelerometer is tilted back 90 degrees over x") {
         mock.setReturnValues(Vec3 { 0, 0, GRAVITY }, Vec3 { 0, 0, 0 }, true);
         accel.begin();
         
-        REQUIRE( accel.getIncline() == Approx(3.14 / 2).margin(0.01));
+        REQUIRE( accel.getIncline().toFloat() == Approx(3.14 / 2).margin(0.01));
     }
 
     SECTION("Accelerometer is tilted forward 90 degrees over x") {
         mock.setReturnValues(Vec3 { 0, 0, -GRAVITY }, Vec3 { 0, 0, 0 }, true);
         accel.begin();
         
-        REQUIRE( accel.getIncline() == Approx(-3.14 / 2).margin(0.01));
+        REQUIRE( accel.getIncline().toFloat() == Approx(-3.14 / 2).margin(0.01));
     }
 
     SECTION("Accelerometer is tilted back 45 degrees over x") {
         mock.setReturnValues(Vec3 { 0, (float)((-GRAVITY)/sqrt(2)), (float)((GRAVITY)/sqrt(2)) }, Vec3 { 0, 0, 0 }, true);
         accel.begin();
         
-        REQUIRE( accel.getIncline() == Approx(3.14 / 4).margin(0.01));
+        REQUIRE( accel.getIncline().toFloat() == Approx(3.14 / 4).margin(0.01));
     }
 
     SECTION("Accelerometer is tilted forward 45 degrees over x") {
         mock.setReturnValues(Vec3 { 0, (float)((-GRAVITY)/sqrt(2)), (float)((-GRAVITY)/sqrt(2)) }, Vec3 { 0, 0, 0 }, true);
         accel.begin();
         
-        REQUIRE( accel.getIncline() == Approx(-3.14 / 4).margin(0.01));
+        REQUIRE( accel.getIncline().toFloat() == Approx(-3.14 / 4).margin(0.01));
     }
 }
